Fix broken test
From what I can tell, assert_not_called doesn't exist in mock.py and
doesn't do anything. Renaming the call to "assert_not_called_1234" also has the tests pass without giving an error like "method doesn't exist".

So replace the non existent method with a check on the number of calls to mock_calculate_hash.

I think there should be 1 call after the comment "hash_path and file_path exist, and have different hashes. This first call will invoke a fetch" and then none after, so I'm not sure why "assert_not_call" was wanted in that location initially as that would be checking that we never were calculating a hash, but the comments suggest we want to check that it doesn't calculate a hash second time?


Bug: chromium:40942322
Change-Id: Idf4bee5d6ee018d47add8704e83a9154293a488c