Fix base_config_unittest test.
This test currently doesn't do anything as assert_call_args doesn't do anything and can be replaced with "assert_call_args_DOESNOTEXIST" and the test passes. This is probably similar to the failure in https://chromium-review.googlesource.com/c/catapult/+/5640871

I tried to create a real cloud_storage_info object to check, but assert_has_calls seems to complain, even when the printout for the contents of the objects are identical (even with any_order=True).



Bug: chromium:40942322
Change-Id: Id8877677e0e82aa0181659373ed6f96b09177cf2