Remove local mock.py from catapult in favor of native mock
Mock has been added to unittest after python 3.3 so it'd be better to use that one.

Most of this change just changes the imports, however some calls in "mock_calls.py" such as mock.name, were moved to mock._mock_name as those variables were moved from public to private between the versions.

The directory paths for httplib2test_appengine.py seem to indicate they're used for python2, but it's believed to not be supported on our systems anymore and the py2 try builders were removed.


Bug: chromium:40942322
Change-Id: I4f0fc4cbf3d27589cd455928da49f21833705dcc