[sandwich][pinpoint] block small regressions from entering bisection
Anomaly detection sometimes will forward anomalies with large expected
differences (comparison magnitude) to regression verification.
Regression verification finds these regressions are actually small
but is precise enough that it still finds a regression and forward
to bisection. Bisection will sometimes fail to repro in these spots.

Here is an example Pinpoint job of this:
https://pinpoint-dot-chromeperf.appspot.com/job/13c114e6c10000

The comparison_magnitude says the diff is expected to be about
279 but regression verification

Change-Id: I35366f94c7844eadaba13415c406e8a855a0b298