Fix non-root permissions
Fixes permissions being erroneously wiped when removing the browser's
profile on most non-rooted devices. This is limited to non-rooted
devices because they fully clear application data instead of just
deleting the profile directory.

This was not caught previously because the most problematic
permission (notifications) is only relevant on Android T+, and we only
started running tests on non-rooted devices with T+ relatively
recently.

Bug: chromium:349132647
Change-Id: I88b55cb4ed217faf551b1902c80e18ca499808ab