[pinpoint] trigger skia culprit finder workflow on confirmed regressions
We are swapping the call from skia bisect to skia culprit finder to
test the E2E workflow. We test on confirmed regressions first and will
eventually test on all potential regressions. After a week or two,
we should confirm that:
- regression verification is able to find regressions
- bisection finds culprits at about a 90% rate
- culprit verification can verify or unverify culprits.

Note that culprit finder is also known as sandwich verification

Bug: chromium:352631795
Change-Id: I9360a4a181f63c578b855e46773ef9809d5e997b