[pinpoint] trigger culprit verification on candidate regressions
We are currently triggering culprit verification in skia only on
confirmed regressions. After ~2 days we have 19/20 culprit finder
workflows in skia that also confirm the regression. The one that did
not failed due to an issue that is now fixed.

We can likely try on all candidate regressions and see if we repro
on non-regressions.

Bug: chromium:352631795
Change-Id: I620d473dfb106b58bb613fdf7d55eb05167170dd