Stop all Android browsers on browser start
Updates browser start on Android to first stop all packages with
"chrome" in the name. This avoids an issue where Telemetry can end up
connecting to the wrong browser instance because another browser is
open on the device and has the `localabstract:chrome_devtools_remote`
port bound already. This case was caught before, but simply failed with
an error that the package that devtools was connected to was incorrect.

In practice, this is not an issue on the bots or on dedicated testing
devices since there is typically no reason to be using multiple browser
APKs at the same time. However, this is very easy to hit if running
tests on a personal device, where it is likely that the device still
has stable Chrome open.

Change-Id: I163ba289a05d9f8ee211943bae6fbba2aab2b18b