Remove defaults property
Removes the defaults property from the optparse/argparse migration
class. It appears that this is only used in one place, so we switch to
directly accessing the protected member in that case. This would ideally
be refactored to avoid relying on default values like this entirely, but
doing so looks like it will require a very large refactor of Telemetry's
argument parsing, which is out of scope for this current work.

Bug: chromium:40807291
Change-Id: I7d0da82333bbfcb0db30df9d52bbd30c4f1007a9