telemetry: Add an ignored '--coverage-dir' arg to run_tests.py
Right now, Chrome's CI unwisely appends the '--coverage-dir' arg to
all android tests that are "generated wrappers". But not all of Chrome's
tests support that arg. telemetry_perf_unittests is one such test. So
migrating telemetry_perf_unittests to generated wrappers causes
it to fail with the following:
no such option: --coverage-dir

To unblock migrating it generated wrappers, this registers that arg
in telemetry's run_tests.py, but ignores it completely. A proper fix
(encoded via TODO here) would involve making Chrome's build smarter
so that it only adds that code-coverage arg for harnesses that
support it.

Bug: chromium:342602628, chromium:368087195
Change-Id: I2091580efec160eaa1cba271f22b1c8cc1206923