[wprgo] Record/replay response stream coming in chunks
This is a prototype. Doesn't even pass test.

Before this CL, wprgo stores the entire response from the server, and
sends that down to the client in one step.

However, some websites have optimized to stream the response as they
process the dynamic contents in the backend.

This CL makes wprgo record/replay the HTTP response streamed to the
client. We hope that the change makes the simulated load characteristics
much closer to the live load.

Change-Id: I473314cf32bdea9d1227d3f519965b1160a45b14