telemetry: Add --assume-browser-already-installed
On android PGO we call run_benchmark over 20 times to produce each
profile, and running UpdateExecutableIfNeeded takes 1+ minute each time
(as on Android there's no faster way to quickly check if a bundle is
already fully installed or not).

Allow callers to bypass UpdateExecutableIfNeeded so when the browser is
known to be installed, the calls can complete much faster.

Bug: chromium:374787804
Change-Id: I4d24692f236c40f63f0579b99ac089de3372bb14