[chromeperf] add helper function to find masters and internal_only by skia host
Now we will load a list of anomalies from Chromeperf and show it on Skia. We need to to filter the anomalies depending on the Skia instance.

This CL adds a helper function that will do the reverse mapping that: given the host name, the function will return the masters and the internal_only value. Then we will use those values to filter the anomalies.

Bug: b:372156352
Change-Id: If95f7cd22ce72b678af93b4e26d37701e522d6f0